requirements.txt
requests
kalshi==0.2.0
python-telegram-bot
pandas
numpy
dotenv

# News and Sentiment Analysis
newsapi-python
transformers
torch
nltk
textblob
requests-cache

# Statistical Arbitrage
statsmodels
scipy
scikit-learn

# Volatility Analysis
arch

# Testing
pytest
pytest-mock